got rid of the editor container, it doesn't do anything meaningful anymore
